Rearrange graphs.ts and utils.ts
