--- data/725a8670-d012-474d-b901-1179f43e7182.json:Zone.Identifier
@@ -1,3 +1,0 @@
-[ZoneTransfer]
-ZoneId=3
-HostUrl=http://pvsite-development-api-sites.eu-west-1.elasticbeanstalk.com/sites/pv_forecast/725a8670-d012-474d-b901-1179f43e7182
